Revert "refactor create-random-id into utils package"
This reverts commit b6c1c71a02caa1e7f39a2bbc0247863570db22b5.
