+++ package.json
@@ -49,7 +49,6 @@
 		"@automattic/wpcom-block-editor": "file:apps/wpcom-block-editor",
 		"browserslist": "4.8.6",
 		"@automattic/calypso-analytics": "file:packages/calypso-analytics",
-		"@automattic/calypso-utils": "file:packages/calypso-utils",
 		"i18n-calypso": "file:packages/i18n-calypso",
 		"photon": "file:packages/photon",
 		"wp-calypso": "file:client",
