+++ packages/calypso-analytics/src/index.js
@@ -10,4 +10,5 @@ export {
 	initializeAnalytics,
 	getTracksAnonymousUserId,
 	analyticsEvents,
+	createRandomId,
 } from './tracks';
