+++ packages/calypso-analytics/src/tracks.js
@@ -1,13 +1,13 @@
 /**
  * External dependencies
  */
-import { assign, includes, isObjectLike, isUndefined, omit, pickBy } from 'lodash';
+import { assign, includes, isObjectLike, isUndefined, omit, pickBy, times } from 'lodash';
 import { parse } from 'qs';
 import cookie from 'cookie';
 import url from 'url';
 import { EventEmitter } from 'events';
 import { loadScript } from '@automattic/load-script';
-import { createRandomId } from '@automattic/calypso-utils';
+
 /**
  * Internal Dependencies
  */
@@ -80,6 +80,21 @@ function checkForBlockedTracks() {
 }
 
 export const analyticsEvents = new EventEmitter();
+
+export function createRandomId( randomBytesLength = 9 ) {
+	// 9 * 4/3 = 12
+	// this is to avoid getting padding of a random byte string when it is base64 encoded
+	let randomBytes;
+
+	if ( window.crypto && window.crypto.getRandomValues ) {
+		randomBytes = new Uint8Array( randomBytesLength );
+		window.crypto.getRandomValues( randomBytes );
+	} else {
+		randomBytes = times( randomBytesLength, () => Math.floor( Math.random() * 256 ) );
+	}
+
+	return window.btoa( String.fromCharCode.apply( String, randomBytes ) );
+}
 
 /**
  * Returns the anoymous id stored in the `tk_ai` cookie
