--- packages/calypso-utils/README.md
@@ -1,4 +1,0 @@
-Utils
-================
-
-A home for simple utility functions that aren't complex enough to justify their own package.
