--- packages/calypso-utils/package.json
@@ -1,35 +1,0 @@
-{
-	"name": "@automattic/utils",
-	"version": "1.0.0-alpha.0",
-	"description": "Automattic utils",
-	"homepage": "https://github.com/Automattic/wp-calypso",
-	"license": "GPL-2.0-or-later",
-	"author": "Automattic Inc.",
-	"sideEffects": false,
-	"main": "dist/cjs/index.js",
-	"module": "dist/esm/index.js",
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/Automattic/wp-calypso.git",
-		"directory": "packages/calypso-utils"
-	},
-	"files": [
-		"dist",
-		"types",
-		"src"
-	],
-	"publishConfig": {
-		"access": "public"
-	},
-	"bugs": {
-		"url": "https://github.com/Automattic/wp-calypso/issues"
-	},
-	"scripts": {
-		"clean": "npx rimraf dist",
-		"prepublish": "npm run clean",
-		"prepare": "transpile"
-	},
-	"dependencies": {
-		"lodash": "^4.17.15"
-	}
-}
