--- packages/calypso-utils/src/index.js
@@ -1,1 +1,0 @@
-export { createRandomId } from './create-random-id';
